Renaming from AggregateRootStateManager to AggregateRootStateProvider
