+++ Specifications/Clients/DotNET/Aggregates/for_AggregateRootStateProvider/when_providing_an_aggregate_without_state_provider.cs
@@ -1,9 +1,9 @@
 // Copyright (c) Aksio Insurtech. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Aksio.Cratis.Aggregates.for_AggregateRootStateManager;
+namespace Aksio.Cratis.Aggregates.for_AggregateRootStateProvider;
 
-public class when_handling_an_aggregate_without_state_provider : given.an_aggregate_root_state_manager
+public class when_providing_an_aggregate_without_state_provider : given.an_aggregate_root_state_manager
 {
     Exception result;
 
@@ -13,7 +13,7 @@ public class when_handling_an_aggregate_without_state_provider : given.an_aggreg
         immediate_projections.Setup(_ => _.HasProjectionFor(typeof(StateForAggregateRoot))).Returns(false);
     }
 
-    async Task Because() => result = await Catch.Exception(() => manager.Handle(aggregate_root, Enumerable.Empty<AppendedEvent>()));
+    async Task Because() => result = await Catch.Exception(() => manager.Provide(aggregate_root, Enumerable.Empty<AppendedEvent>()));
 
     [Fact] void should_throw_missing_state_provider_exception() => result.ShouldBeOfExactType<MissingAggregateRootStateProvider>();
 }
